move free to free folder
