+++ alloc/alloc.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <flip_social.h>
 #include <callback/callback.h>
-#include <alloc/free.h>
+#include <free/free.h>
 void auth_headers_alloc(void);
 FlipSocialFeedMini *flip_feed_info_alloc(void);
 bool messages_dialog_alloc(bool free_first);
