+++ callback/callback.h
@@ -7,7 +7,7 @@
 #include <explore/flip_social_explore.h>
 #include <feed/flip_social_feed.h>
 #include <flip_storage/flip_social_storage.h>
-#include <alloc/free.h>
+#include <free/free.h>
 #include <alloc/alloc.h>
 
 /**
